rename the math.h to be mymath.h to avoid the same name as system's since it's failted to be built on linux
