+++ src/step10/rayweek1.cpp
@@ -31,7 +31,7 @@
 #include <immintrin.h>
 
 #include "../common/common.h"
-#include "math.h"
+#include "mymath.h"
 #include "soa_sphere.h"
 
 #define SIMD_WIDTH  4
