+++ src/step13/rayweek1.cpp
@@ -32,7 +32,7 @@
 #define MULTITHREADED   1
 
 #include "../common/common.h"
-#include "math.h"
+#include "mymath.h"
 #include "soa_sphere.h"
 
 #define SIMD_WIDTH  8
