+++ src/step7/mymath.h
